A sample .butanerc which goes in the user's home diretcory.
It should work with multiple campfire accounts.  In the sample, there is a
single campfire account configured, mycampfire.

My login and password are provided and then a section for multiple rooms to be
monitored in the "mycampfire" account.

There is the "First Room".  Any activity that regexp matches /rick/i will pop
up a sticky notification that needs to be clicked away.

In the "Git Commits" room, any activity that matches the regexp /models|specs/i
will cause a sticky notification to appear.

Any other traffic in either of those rooms which pop up a notification that
disappears after 5 seconds.

The intent is that multiple accounts can be configured each with one or more
rooms that will be monitored for traffic.