Hmmmm ... the new tinder gem uses eventmachine and I'm not clear yet on the interaction between the threads that butane fires up and EM.  After all rooms, each their own thread, have started monitoring, all the threads quit save for the first room thread.
There are only 2 threads that run.  The main thread and the first room thread.
But, all the rooms remain monitored even after the associated room thread
quits.

So, this message about 'stopped monitoring ...' is inaccurate.

I'm not yet sure on how to know if a room is no longer being monitored tho.

I guess I need to spend a bit of time with EM and threads. :)