Move to using EventMachine and ruby 1.9.2
The best way to use the new butane is with rvm 1.9.2 and the bundle gem.

The 'tinder' gem now uses EventMachine and this reflects part of that
change. It looks like an individual Tinder::Room instance will fire up
and EM reactor but what is needed here is a method to listen to multiple
rooms.

I didn't see anything like that, so I added a class method called
listen_to_rooms to handle that. The majority of the code comes directly
from the Tinder::Room#listen method but loops through an array of rooms
inside of a an EM reactor loop.

Threads are no longer used (well, except for the extra thread that gets
created by EM)